requirements.txt
python-telegram-bot==20.7
nextcord==2.5.0